directories, subclass ideas, new shape term
Split everything into directories for datasets, source code, and header files. Added some comments about class inheritance for single-seat maps (for when the population term is generalised for multiple seats) and no county borders (for when county borders are considered). Spitballed an idea about combining contiguity and compactness into a single term.

+++ headers/Map.cpp
@@ -78,7 +78,7 @@ valarray<double> Map::deltaH_prop_(const int& x, const int& prop, vector<int>& p
   int deltaHD_prop = 0;
   for (int i = 0; i < ED_nei_[x].size(); i ++) if (ED_q_[ED_nei_[x][i]] == prop) deltaHD_prop --;
 
-  return valarray<double>{ (abs(q_pop_[prop] + ED_pop_[x]) - abs(q_pop_[prop]))/av_pop_, q_group_[prop].size() - pngs.size() - fabs(q_group_[prop].size() - 1.), double(deltaHD_prop) };
+  return valarray<double>{ double(abs(q_pop_[prop] + ED_pop_[x]) - abs(q_pop_[prop]))/av_pop_, q_group_[prop].size() - pngs.size() - fabs(q_group_[prop].size() - 1.), double(deltaHD_prop) };
 }
 
 void Map::config_update_()
